Rearranged JS
