+++ scripts/common.js
@@ -1,23 +1,8 @@
-/** Bootstrap Form Validation */
-(() => {
-  "use strict";
+import { handleLogOut } from "./utils/handleLogOut.mjs";
+import { bootstrapFormValidation } from "./utils/bootstrapFormValidation.mjs";
 
-  // Fetch all the forms we want to apply custom Bootstrap validation styles to
-  const forms = document.querySelectorAll(".needs-validation");
-
-  // Loop over them and prevent submission
-  Array.from(forms).forEach((form) => {
-    form.addEventListener(
-      "submit",
-      (event) => {
-        if (!form.checkValidity()) {
-          event.preventDefault();
-          event.stopPropagation();
-        }
-
-        form.classList.add("was-validated");
-      },
-      false
-    );
-  });
-})();
+function app() {
+  bootstrapFormValidation();
+  handleLogOut();
+}
+app();
